Reduce MTF converter maintenance risk
Extract serialized DTO definitions from the MegaMekLab MTF converter so the converter drops below the high Python file-bloat threshold while preserving its CLI and output model.

Constraint: maintenance-code-health remains active-review until remaining repo-wide critical/high/warn ledger debt is burned down or explicitly accepted.

Rejected: promote maintenance-code-health to ready-with-scope in this slice | remaining file-bloat and near-duplicate follow-ups are still live.

Confidence: high

Scope-risk: narrow

Directive: Keep future converter splits behavior-preserving and backed by converter fixture parity before changing parsing rules.

Tested: python -B scripts/megameklab-conversion/mtf_converter.py --help; npm.cmd run maintain:scan -- --scope=scripts/megameklab-conversion/mtf_converter.py --json; npm.cmd run maintain:scan -- --scope=scripts/megameklab-conversion/mtf_serialized_models.py --json; node scripts/qc/validate-maintenance-warning-ledger.mjs; npm.cmd run qc:validate; npm.cmd run verify:qc:maintenance; npm.cmd run qc:app-completion -- --json; git diff --check

Not-tested: full MegaMekLab fixture conversion; full verify:app-completion remains blocked by maintenance-code-health active-review.
