+++ docs/qc/maintenance-warning-ledger.json
@@ -83,8 +83,18 @@
       "severity": "high",
       "file": "scripts/megameklab-conversion/mtf_converter.py",
       "message": "546 LOC exceeds python threshold",
+      "status": "fixed",
+      "rationale": "Serialized DTO dataclasses were extracted to mtf_serialized_models.py, dropping mtf_converter.py below the high Python file-bloat threshold while preserving converter behavior.",
+      "followUps": ["wave-12-megameklab-conversion-split"]
+    },
+    {
+      "key": "file-bloat|warn|scripts/megameklab-conversion/mtf_converter.py|487 LOC exceeds python threshold",
+      "category": "file-bloat",
+      "severity": "warn",
+      "file": "scripts/megameklab-conversion/mtf_converter.py",
+      "message": "487 LOC exceeds python threshold",
       "status": "follow-up",
-      "rationale": "MegaMekLab MTF conversion should be split with converter fixture coverage rather than changed inside this ledger wave.",
+      "rationale": "The MTF converter remains above the advisory Python size threshold after DTO extraction; finish the split only with converter fixture parity coverage.",
       "followUps": ["wave-12-megameklab-conversion-split"]
     },
     {
