+++ scripts/megameklab-conversion/mtf_converter.py
@@ -17,7 +17,6 @@ import re
 import argparse
 from pathlib import Path
 from typing import Dict, List, Optional, Any, Tuple
-from dataclasses import dataclass, asdict
 from enum_mappings import (
     map_tech_base,
     map_rules_level,
@@ -36,102 +35,17 @@ from enum_mappings import (
     get_era_folder_name,
     get_rules_level_folder_name,
 )
-
-
-@dataclass
-class SerializedEngine:
-    """Engine configuration."""
-    type: str
-    rating: int
-
-
-@dataclass
-class SerializedGyro:
-    """Gyro configuration."""
-    type: str
-
-
-@dataclass
-class SerializedStructure:
-    """Internal structure configuration."""
-    type: str
-
-
-@dataclass
-class SerializedArmor:
-    """Armor configuration."""
-    type: str
-    allocation: Dict[str, Any]  # Can be int or {front: int, rear: int}
-
-
-@dataclass
-class SerializedHeatSinks:
-    """Heat sink configuration."""
-    type: str
-    count: int
-
-
-@dataclass
-class SerializedMovement:
-    """Movement configuration."""
-    walk: int
-    jump: int
-    jumpJetType: Optional[str] = None
-    enhancements: Optional[List[str]] = None
-
-
-@dataclass
-class SerializedEquipment:
-    """Mounted equipment item."""
-    id: str
-    location: str
-    slots: Optional[List[int]] = None
-    isRearMounted: Optional[bool] = None
-    linkedAmmo: Optional[str] = None
-
-
-@dataclass
-class SerializedFluff:
-    """Fluff/flavor text."""
-    overview: Optional[str] = None
-    capabilities: Optional[str] = None
-    history: Optional[str] = None
-    deployment: Optional[str] = None
-    variants: Optional[str] = None
-    notableUnits: Optional[str] = None
-    manufacturer: Optional[str] = None
-    primaryFactory: Optional[str] = None
-    systemManufacturer: Optional[Dict[str, str]] = None
-
-
-@dataclass
-class SerializedUnit:
-    """Complete serialized unit matching ISerializedUnit interface."""
-    id: str
-    chassis: str
-    model: str
-    unitType: str
-    configuration: str
-    techBase: str
-    rulesLevel: str
-    era: str
-    year: int
-    tonnage: int
-    engine: SerializedEngine
-    gyro: SerializedGyro
-    cockpit: str
-    structure: SerializedStructure
-    armor: SerializedArmor
-    heatSinks: SerializedHeatSinks
-    movement: SerializedMovement
-    equipment: List[SerializedEquipment]
-    criticalSlots: Dict[str, List[Optional[str]]]
-    variant: Optional[str] = None
-    quirks: Optional[List[str]] = None
-    fluff: Optional[SerializedFluff] = None
-    mulId: Optional[int] = None
-    role: Optional[str] = None
-    source: Optional[str] = None
+from mtf_serialized_models import (
+    SerializedArmor,
+    SerializedEngine,
+    SerializedEquipment,
+    SerializedFluff,
+    SerializedGyro,
+    SerializedHeatSinks,
+    SerializedMovement,
+    SerializedStructure,
+    SerializedUnit,
+)
 
 
 class MTFParser:
@@ -754,4 +668,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
